@@ -1,11 +1,13 @@
 {
   "name": "fork-ts-checker-notifier-webpack-plugin",
-  "version": "6.0.0",
+  "version": "7.0.0",
   "description": "a notifier for users of fork-ts-checker-webpack-plugin",
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
     "build": "tsc --version && tsc",
+    "format": "prettier --check .",
+    "lint": "eslint --color --ext js,ts .",
     "test": "jest"
   },
   "keywords": [
@@ -15,7 +17,7 @@
     "notifier"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
   "author": "John Reilly <johnny_reilly@hotmail.com>",
   "license": "MIT",
@@ -27,22 +29,26 @@
     "url": "https://github.com/johnnyreilly/fork-ts-checker-notifier-webpack-plugin/issues"
   },
   "dependencies": {
-    "node-notifier": "^8.0.2"
+    "node-notifier": "^10.0.1"
   },
   "devDependencies": {
-    "@types/jest": "^23.3.2",
-    "@types/node": "^17.0.9",
-    "@types/node-notifier": "^0.0.28",
-    "@types/tapable": "^1.0.4",
-    "fork-ts-checker-webpack-plugin": "^7.2.1",
-    "jest": "^27.4.7",
-    "ts-jest": "^27.1.2",
-    "tslint": "^5.12.0",
-    "typescript": "^4.5.4",
-    "webpack": "^5.69.1"
+    "@types/jest": "^29.5.1",
+    "@types/node": "^18.16.0",
+    "@types/node-notifier": "^8.0.2",
+    "@types/tapable": "^2.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "fork-ts-checker-webpack-plugin": "^8.0.0",
+    "jest": "^29.5.0",
+    "prettier": "^2.8.8",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4",
+    "webpack": "^5.80.0"
   },
   "peerDependencies": {
-    "fork-ts-checker-webpack-plugin": "^7.0.0"
+    "fork-ts-checker-webpack-plugin": "^8.0.0"
   },
   "jest": {
     "transform": {
