@@ -1,7 +1,5 @@
-import ForkTsCheckerNotifierWebpackPlugin = require('./index');
-import {
-  Issue
-} from 'fork-ts-checker-webpack-plugin/lib/issue';
+import ForkTsCheckerNotifierWebpackPlugin from './index';
+import { Issue } from 'fork-ts-checker-webpack-plugin/lib/issue';
 
 const error: Issue = {
   code: 'code',
@@ -11,13 +9,13 @@ const error: Issue = {
   location: {
     start: {
       line: 10,
-      column: 10
+      column: 10,
     },
     end: {
       line: 10,
-      column: 10
-    }
-  }
+      column: 10,
+    },
+  },
 };
 
 const warning: Issue = {
@@ -28,19 +26,19 @@ const warning: Issue = {
   location: {
     start: {
       line: 10,
-      column: 10
+      column: 10,
     },
     end: {
       line: 10,
-      column: 10
-    }
-  }
+      column: 10,
+    },
+  },
 };
 
 describe('buildNotification', () => {
   test('first notification is skipped', () => {
     const plugin = new ForkTsCheckerNotifierWebpackPlugin({
-      skipFirstNotification: true
+      skipFirstNotification: true,
     });
 
     const returnValue = plugin.buildNotification([]);
@@ -54,9 +52,13 @@ describe('buildNotification', () => {
 
     const returnValue = plugin.buildNotification([warning, error]);
 
-    expect(returnValue!.title).toBe(`Error in ${error.file}`);
-    expect(returnValue!.message).toBe(error.message);
-    expect(returnValue!.icon.endsWith('error.png')).toBe(true);
+    if (returnValue === undefined) {
+      throw Error();
+    }
+
+    expect(returnValue.title).toBe(`Error in ${error.file}`);
+    expect(returnValue.message).toBe(error.message);
+    expect(returnValue.icon.endsWith('error.png')).toBe(true);
     expect(plugin.lastBuildSucceeded).toBe(false);
   });
 
@@ -65,37 +67,49 @@ describe('buildNotification', () => {
 
     const returnValue = plugin.buildNotification([warning]);
 
-    expect(returnValue!.title).toBe(`Warning in ${warning.file}`);
-    expect(returnValue!.message).toBe(warning.message);
-    expect(returnValue!.icon.endsWith('warning.png')).toBe(true);
+    if (returnValue === undefined) {
+      throw Error();
+    }
+
+    expect(returnValue.title).toBe(`Warning in ${warning.file}`);
+    expect(returnValue.message).toBe(warning.message);
+    expect(returnValue.icon.endsWith('warning.png')).toBe(true);
     expect(plugin.lastBuildSucceeded).toBe(false);
   });
 
   test('given warning when excludeWarnings is true should return success', () => {
     const plugin = new ForkTsCheckerNotifierWebpackPlugin({
-      excludeWarnings: true
+      excludeWarnings: true,
     });
 
     const returnValue = plugin.buildNotification([warning]);
 
-    expect(returnValue!.title).toBe('Type check succeeded');
-    expect(returnValue!.message).toBe(
+    if (returnValue === undefined) {
+      throw Error();
+    }
+
+    expect(returnValue.title).toBe('Type check succeeded');
+    expect(returnValue.message).toBe(
       'No type errors! See warning(s) in console!'
     );
-    expect(returnValue!.icon.endsWith('built.png')).toBe(true);
+    expect(returnValue.icon.endsWith('built.png')).toBe(true);
     expect(plugin.lastBuildSucceeded).toBe(true);
   });
 
   test('given no errors or warnings should return success', () => {
     const plugin = new ForkTsCheckerNotifierWebpackPlugin({
-      excludeWarnings: true
+      excludeWarnings: true,
     });
 
     const returnValue = plugin.buildNotification([]);
 
-    expect(returnValue!.title).toBe('Type check succeeded');
-    expect(returnValue!.message).toBe('No type errors!');
-    expect(returnValue!.icon.endsWith('built.png')).toBe(true);
+    if (returnValue === undefined) {
+      throw Error();
+    }
+
+    expect(returnValue.title).toBe('Type check succeeded');
+    expect(returnValue.message).toBe('No type errors!');
+    expect(returnValue.icon.endsWith('built.png')).toBe(true);
     expect(plugin.lastBuildSucceeded).toBe(true);
   });
 });
